backend/requirements.txt
django==4.2.1
django-stdimage==6.0.1
gunicorn==20.1.0
sqlparse==0.4.4
whitenoise==6.4.0